Address a bunch of style and other comments
