--- resource-managers/kubernetes/core/src/main/scala/org/apache/spark/deploy/k8s/OptionRequirements.scala
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.spark.deploy.k8s
-
-private[spark] object OptionRequirements {
-
-  def requireBothOrNeitherDefined(
-      opt1: Option[_],
-      opt2: Option[_],
-      errMessageWhenFirstIsMissing: String,
-      errMessageWhenSecondIsMissing: String): Unit = {
-    requireSecondIfFirstIsDefined(opt1, opt2, errMessageWhenSecondIsMissing)
-    requireSecondIfFirstIsDefined(opt2, opt1, errMessageWhenFirstIsMissing)
-  }
-
-  def requireSecondIfFirstIsDefined(
-      opt1: Option[_], opt2: Option[_], errMessageWhenSecondIsMissing: String): Unit = {
-    opt1.foreach { _ =>
-      require(opt2.isDefined, errMessageWhenSecondIsMissing)
-    }
-  }
-
-  def requireNandDefined(opt1: Option[_], opt2: Option[_], errMessage: String): Unit = {
-    opt1.foreach { _ => require(opt2.isEmpty, errMessage) }
-  }
-}
